Change the behavior of updating and removing directory in settings.
* When attempting to remove a directory, it will fail and warn users if there are some installed collection from that repository. Users can remove a directory if there is no installed collections from that directory.
* When attempting to update a directory, it will check if there's installed collection from the old URL. Old installed collection exists in the new URL are identified by its register name. Cases for installed collections:
	1. Old collection exists in the new URL, same URL: use the new one, update the status to INSTALLED
	2. Old collection exists in the new URL, different URL: keep them both (add the old one). Because they should be treated as different collection
	3. Old collection doesn't exist in the new URL, same URL: keep the old collection (adding it to new collections)
	4. Old collection doesn't exist in the new URL, different URL: same with 3.
* For the installed collection that the repository URL is not registered in the settings, once it is uninstalled and users reload the repository, it won't be listed in the list of collections

+++ resource_sharing/gui/resource_sharing_dialog.py
@@ -314,7 +314,12 @@ class ResourceSharingDialog(QtGui.QDialog, FORM_CLASS):
         # Edit repository
         try:
             status, description = self.repository_manager.edit_directory(
-                repo_name, new_name, new_url, new_auth_cfg)
+                repo_name,
+                new_name,
+                old_url,
+                new_url,
+                new_auth_cfg
+            )
             if status:
                 self.message_bar.pushMessage(
                     self.tr('Repository is successfully updated'),
@@ -345,10 +350,8 @@ class ResourceSharingDialog(QtGui.QDialog, FORM_CLASS):
         if not repo_name:
             return
         # Check if it's the approved online dir repository
-        settings = QSettings()
-        settings.beginGroup(repo_settings_group())
-        if settings.value(repo_name + '/url') in \
-                self.repository_manager._online_directories.values():
+        repo_url = self.repository_manager.directories[repo_name]['url']
+        if repo_url in self.repository_manager._online_directories.values():
             self.message_bar.pushMessage(
                 self.tr(
                     'You can not remove the official repositories!'),
@@ -366,14 +369,19 @@ class ResourceSharingDialog(QtGui.QDialog, FORM_CLASS):
             return
 
         # Remove repository
-        self.repository_manager.remove_directory(repo_name)
-
-        # Reload data and widget
-        self.reload_data_and_widget()
-
-        # Deactivate edit and delete button
-        self.button_edit.setEnabled(False)
-        self.button_delete.setEnabled(False)
+        installed_collections = \
+            self.collection_manager.get_installed_collections(repo_url)
+        if installed_collections:
+            message = ('You have some installed collections from this '
+                       'repository. Please uninstall them first!')
+            self.message_bar.pushMessage(message, QgsMessageBar.WARNING, 5)
+        else:
+            self.repository_manager.remove_directory(repo_name)
+            # Reload data and widget
+            self.reload_data_and_widget()
+            # Deactivate edit and delete button
+            self.button_edit.setEnabled(False)
+            self.button_delete.setEnabled(False)
 
     def reload_repositories(self):
         """Slot for when user clicks reload repositories button."""
